api/requirements.txt
django
djangorestframework
pygments
django-cors-headers
pillow
markdown
django-filter
mysqlclient